LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-activation/activation-initial-about-blank-expected.txt



FAIL navigation.activation interaction with initial about:blank assert_equals: expected null but got object "[object NavigationHistoryEntry]"
